--- src/main/java/bumblebee/xchangepass/global/security/login/dto/LoginResponse.java
@@ -1,12 +1,0 @@
-package bumblebee.xchangepass.global.security.login.dto;
-
-import lombok.Builder;
-
-public record LoginResponse(
-        String accessToken,
-        String refreshToken
-) {
-    @Builder
-    public LoginResponse {
-    }
-}
