+++ src/test/java/bumblebee/xchangepass/domain/user/service/UserLoginUnitTest.java
@@ -1,7 +1,7 @@
 package bumblebee.xchangepass.domain.user.service;
 
 import bumblebee.xchangepass.domain.user.dto.request.UserRegisterRequest;
-import bumblebee.xchangepass.domain.user.dto.response.UserLoginResponse;
+import bumblebee.xchangepass.global.security.login.dto.response.UserLoginResponse;
 import bumblebee.xchangepass.domain.user.entity.Role;
 import bumblebee.xchangepass.domain.user.entity.Sex;
 import bumblebee.xchangepass.domain.user.repository.UserRepository;
@@ -9,8 +9,8 @@ import bumblebee.xchangepass.global.exception.CommonException;
 import bumblebee.xchangepass.global.security.jwt.JwtProvider;
 import bumblebee.xchangepass.global.security.login.LoginService;
 import bumblebee.xchangepass.global.security.login.UserRegisterService;
-import bumblebee.xchangepass.global.security.login.dto.LoginRequest;
-import bumblebee.xchangepass.global.security.login.dto.LoginResponse;
+import bumblebee.xchangepass.global.security.login.dto.request.LoginRequest;
+import bumblebee.xchangepass.global.security.login.dto.response.LoginResponse;
 import bumblebee.xchangepass.global.security.refresh.RefreshToken;
 import bumblebee.xchangepass.global.security.refresh.RefreshTokenService;
 import bumblebee.xchangepass.global.security.refresh.dto.RefreshTokenResponse;
